bug/minor: ux: fix issue with container qty/unit not being editable
after creation. make reloadElements also reload malleable stuff.
